move capi search results to their own component
